+++ design_docs/history.md
@@ -15,6 +15,7 @@
   - [17/02/25](#170225)
   - [18/02/25](#180225)
   - [19/02/25](#190225)
+  - [20/02/25](#200225)
 
 ## November 24
 
@@ -104,3 +105,8 @@ harder to focus today
 - Tiles and Grid Refactor : HexGrid Lib 9-10-11-12
 - CSS Animation Synchronisation 14-15h30
 - CSS Animation Circle Animations 15h30-17h
+
+## 20/02/25
+
+- Gameplay System : PlayedAction, ActionSequences 10-12
+- Replay System : ReplayWriter